+++ src/ActiveLogin.Authentication.GrandId.AspNetCore/GrandIdBankIdAuthenticationPostConfigureOptions.cs
@@ -1,37 +1,12 @@
-﻿using ActiveLogin.Authentication.GrandId.AspNetCore.Models;
-using ActiveLogin.Authentication.GrandId.AspNetCore.Serialization;
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.DataProtection;
-using Microsoft.Extensions.Options;
+﻿using Microsoft.AspNetCore.DataProtection;
 
 namespace ActiveLogin.Authentication.GrandId.AspNetCore
 {
-    public class GrandIdBankIdAuthenticationPostConfigureOptions : IPostConfigureOptions<GrandIdBankIdAuthenticationOptions>
+    public class GrandIdBankIdAuthenticationPostConfigureOptions : GrandIdAuthenticationPostConfigureOptions<GrandIdBankIdAuthenticationOptions, GrandIdBankIdAuthenticationHandler>
     {
-        private readonly IDataProtectionProvider _dp;
-
         public GrandIdBankIdAuthenticationPostConfigureOptions(IDataProtectionProvider dataProtection)
+            : base(dataProtection)
         {
-            _dp = dataProtection;
-        }
-
-        public void PostConfigure(string name, GrandIdBankIdAuthenticationOptions options)
-        {
-            options.DataProtectionProvider = options.DataProtectionProvider ?? _dp;
-
-            if (options.StateDataFormat == null)
-            {
-                var dataProtector = options.DataProtectionProvider.CreateProtector(
-                    typeof(GrandIdBankIdAuthenticationHandler).FullName,
-                    name,
-                    "v1"
-                );
-
-                options.StateDataFormat = new SecureDataFormat<GrandIdState>(
-                    new GrandIdStateSerializer(),
-                    dataProtector
-                );
-            }
         }
     }
 }
