+++ src/ActiveLogin.Authentication.GrandId.AspNetCore/GrandIdLoggingEvents.cs
@@ -5,8 +5,8 @@ namespace ActiveLogin.Authentication.GrandId.AspNetCore
     internal static class GrandIdLoggingEvents
     {
         // GrandId API - Auth
-        public static readonly EventId GrandIdAuthSuccess = new EventId(1_1_1, nameof(GrandIdAuthSuccess));
-        public static readonly EventId GrandIdAuthHardFailure = new EventId(1_1_2, nameof(GrandIdAuthHardFailure));
+        public static readonly EventId GrandIdFederatedLoginSuccess = new EventId(1_1_1, nameof(GrandIdFederatedLoginSuccess));
+        public static readonly EventId GrandIdFederatedLoginHardFailure = new EventId(1_1_2, nameof(GrandIdFederatedLoginHardFailure));
 
         // GrandId API - Get session
         public static readonly EventId GrandIdGetSessionSuccess = new EventId(2_1_1, nameof(GrandIdGetSessionSuccess));
